refactor: aftereach로 수정
