+++ src/test/java/roomescape/integration/repository/ReservationTimeRepositoryTestBase.java
@@ -11,7 +11,7 @@ import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import roomescape.common.RepositoryBaseTest;
+import roomescape.common.RepositoryTestBase;
 import roomescape.domain.theme.Theme;
 import roomescape.domain.time.AvailableReservationTime;
 import roomescape.domain.time.ReservationTime;
@@ -21,7 +21,7 @@ import roomescape.integration.fixture.ReservationDbFixture;
 import roomescape.integration.fixture.ReservationTimeDbFixture;
 import roomescape.integration.fixture.ThemeDbFixture;
 
-public class ReservationTimeRepositoryTest extends RepositoryBaseTest {
+public class ReservationTimeRepositoryTestBase extends RepositoryTestBase {
 
     @Autowired
     private ReservationTimeRepository repository;
